New endpoints where back-end servers can register with the api/controller.
Once a server registers, the controller sends it regular pings; if it doesn't respond to a few pings in a row then it gets dropped from the server list.

Related work items: #978

+++ service/src/test/java/com/hutoma/api/common/TestThreadPool.java
@@ -151,7 +151,7 @@ public class TestThreadPool {
         }
     }
 
-    TestableThreadPool getPool(int max, int lifespan) {
+    TestableThreadPool getPool(int max, long lifespan) {
         Config mockConfig = mock(Config.class);
         when(mockConfig.getThreadPoolMaxThreads()).thenReturn(max);
         when(mockConfig.getThreadPoolIdleTimeMs()).thenReturn(lifespan);
